+++ test/testkit/SilhouetteFakeModule.scala
@@ -1,0 +1,28 @@
+/*
+ * Copyright 2017 HM Revenue & Customs
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package testkit
+
+import com.google.inject.AbstractModule
+import com.mohiva.play.silhouette.api.Environment
+import net.codingwell.scalaguice.ScalaModule
+import security.SecurityEnvironment
+
+class SilhouetteFakeModule extends AbstractModule with SilhouetteFakeEnv with ScalaModule {
+  def configure(): Unit = {
+    bind[Environment[SecurityEnvironment]].toInstance(env)
+  }
+}
